Fix for Hynek's comments

+++ adapters/saml/core/src/main/java/org/keycloak/adapters/cloned/AdapterHttpClientConfig.java
@@ -76,11 +76,11 @@ public interface AdapterHttpClientConfig {
     /**
      * Returns timeout for socket waiting for data in milliseconds.
      */
-    int getSocketTimeout();
+    long getSocketTimeout();
 
     /**
      * Returns timeout for establishing the connection with the remote host in milliseconds.
      */
-    int getConnectionTimeout();
+    long getConnectionTimeout();
 
 }
